Split todos and categories code into their own seperate controllers
