+++ src/javascript/toggleMenus.js
@@ -16,40 +16,6 @@ const TodoMenu = () => {
   });
 };
 
-const AddTodoMenu = () => {
-  const addTodoMenuBtns = document.querySelectorAll('.add-todo-btn, .cancel-add-new-todo-button');
-  const addNewTodoMenu = document.querySelector('.add-new-todo-container');
-  const addCategoryBtns = document.querySelectorAll('.category-btn-add');
-
-  addTodoMenuBtns.forEach((addTodoMenuBtn) => {
-    addTodoMenuBtn.addEventListener('click', () => {
-      addNewTodoMenu.classList.toggle('show-add-new-todo');
-    });
-  });
-
-  addCategoryBtns.forEach(((addCategoryBtn) => {
-    addCategoryBtn.addEventListener('click', (e) => {
-      if (!e.currentTarget.classList.contains('active-category')) {
-        addCategoryBtns.forEach((addCategoryBtn) => {
-          addCategoryBtn.classList.remove('active-category');
-        });
-        e.currentTarget.classList.add('active-category');
-      }
-    });
-  }));
-};
-
-const EditTodoMenu = () => {
-  const editTodoMenuBtns = document.querySelectorAll('.todo-edit-btn, .cancel-edit-existing-todo-button');
-  const editExistingTodoMenu = document.querySelector('.edit-existing-todo-container');
-
-  editTodoMenuBtns.forEach((editTodoMenuBtn) => {
-    editTodoMenuBtn.addEventListener('click', () => {
-      editExistingTodoMenu.classList.toggle('show-edit-existing-todo');
-    });
-  });
-};
-
 const DeleteTodoMenu = () => {
   const deleteTodoMenuBtns = document.querySelectorAll('.todo-delete-btn, .cancel-delete-todo-button');
   const deleteTodoMenu = document.querySelector('.delete-todo-popup-container');
@@ -94,26 +60,12 @@ const DeleteCategoryMenu = () => {
   });
 };
 
-const AddCategoryMenu = () => {
-  const addCategoryMenuBtns = document.querySelectorAll('.add-category-button, .cancel-add-category-button');
-  const addCategoryMenu = document.querySelector('.add-category-popup-container');
-
-  addCategoryMenuBtns.forEach((addCategoryMenuBtn) => {
-    addCategoryMenuBtn.addEventListener('click', () => {
-      addCategoryMenu.classList.toggle('show-add-category-popup');
-    });
-  });
-};
-
 const setMenus = () => {
   TodoMenu();
-  AddTodoMenu();
-  EditTodoMenu();
   DeleteTodoMenu();
   DeleteAllTodoMenu();
   EditCategoriesMenu();
   DeleteCategoryMenu();
-  AddCategoryMenu();
 };
 
 export default setMenus;
